Added controller for logs
