+++ routes/Route.js
@@ -3,8 +3,8 @@ const router = express.Router('router');
 const User = require('../models/user.js');
 const bcrypt = require('bcrypt');
 const jwt = require('jsonwebtoken');
-
 const userController = require('../controller/userController');
+const logController = require('../controller/logController');
 require('dotenv').config();
 
 
@@ -15,77 +15,14 @@ router.get('/', (req, res)=>{
 });
 
 
-router.post('/userExists', userController.userExists);  //username exists or not
-router.get('/verify/:authToken',userController.verifyToken);  // verify the auth token
-router.post('/signup', userController.signup);
-router.post('/login', userController.login);
-router.post('/displayLogs', userController.displayLogs);
+router.post('/userExists'       , userController.userExists);  //username exists or not
+router.get ('/verify/:authToken',userController.verifyToken);  // verify the auth token
+router.post('/signup'           , userController.signup);
+router.post('/login'            , userController.login);
+router.post('/displayLogs'      , logController.displayLogs);
+router.post('/createLog'        , logController.createLog); //done - handled if no user is matched, if no new loglength is 0
 
-//done - handled if no user is matched, if no new loglength is 0
-router.post('/createLog', (req, res) => {
-
-    var authToken = req.header("Authorization");
-    authToken = authToken.substr(7, authToken.length);
-    console.log(authToken);
-    jwt.verify(authToken, process.env.JWT_ACC_ACTIVATE1, (error, decodedToken) => {
-        if (error) {
-            console.log(error);
-            return res.status(500).json({
-                error: 'Incorrect or Expired Link.'
-            });
-        }
-        else {
-            console.log("print log");
-            const { email, username } = decodedToken;
-            console.log(email);
-            console.log(username);
-
-
-            //var username = req.body.username;
-            var newlog = req.body.text;
-            var timeNow = new Date();
-
-            console.log(newlog);
-
-            if (newlog != undefined && newlog.length) {
-                User.update({ username: username }, { $push: { "logs": [{ "timestamp": timeNow, "logs": newlog }] } })
-                    .exec()
-                    .then(result => {
-                        if (result.matchedCount) {
-                            console.log(result);
-                            res.status(200).json({
-                                message: "log updated for" + username
-                            });
-                        }
-                        else {
-                            console.log("No user as " + username + " found");
-                            res.status(200).json({
-                                message: "No user as " + username + " found"
-                            });
-                        }
-                    })
-                    .catch(error => {
-                        console.log(error);
-                        res.status(500).json({
-                            message: "something bad happened.",
-                            error: error
-                        });
-                    });
-            }
-            else {
-                console.log("No log found. Please enter a log.");
-                res.status(200).json({
-                    message: "No log found. Please enter a log."
-                });
-            }
-        }
-    });
-
-
-
-});
-
-//done 
+//done  
 router.post('/setThought', (req, res) => {
 
     var username = req.body.username;
